Redistribute auxilliry methods by matric provider class, rather than execution engine
