+++ pyproject.toml
@@ -156,9 +156,10 @@ exclude = [
     'expectations/metrics/map_metric_provider/multi_column_map_metric_provider\.py', # 2
     'expectations/metrics/map_metric_provider/multicolumn_condition_partial\.py', # 3
     'expectations/metrics/map_metric_provider/multicolumn_function_partial\.py', # 2
-    'expectations/metrics/map_metric_provider/pandas_methods\.py', # 7
-    'expectations/metrics/map_metric_provider/spark_methods\.py', # 10
-    'expectations/metrics/map_metric_provider/sqlalchemy_methods\.py', # 20
+	'expectations/metrics/map_metric_provider/column_map_condition_auxilliary_methods\.py',
+	'expectations/metrics/map_metric_provider/column_pair_map_condition_auxilliary_methods\.py',
+	'expectations/metrics/map_metric_provider/map_condition_auxilliary_methods\.py',
+	'expectations/metrics/map_metric_provider/multicolumn_map_condition_auxilliary_methods\.py',
     'expectations/regex_based_column_map_expectation\.py',  # 3
     'expectations/row_conditions\.py',  # 4
     'expectations/set_based_column_map_expectation\.py',  # 3
